+++ src/divide.c
@@ -34,12 +34,14 @@ typedef enum {
 
 static const div_rule RULES[FMT_N_FIELDS] = {
     FMT_CHANNELS(CHANNEL)
-    [FMT_COVERAGE] = DIV_SUM,
-    [FMT_LENGTHS]  = DIV_SUM,
-    [FMT_READS]    = DIV_SUM,
-    [FMT_REJECTED] = DIV_SUM,
-    [FMT_UNMAPPED] = DIV_SUM,
-    [FMT_SEQUENCE] = DIV_SAME,
+    [FMT_COVERAGE]                 = DIV_SUM,
+    [FMT_LENGTHS]                  = DIV_SUM,
+    [FMT_PRIMARY_COUNTED]          = DIV_SUM,
+    [FMT_PRIMARY_REJECTED]         = DIV_SUM,
+    [FMT_SUPPLEMENTARY_COUNTED]    = DIV_SUM,
+    [FMT_SUPPLEMENTARY_REJECTED]   = DIV_SUM,
+    [FMT_UNMAPPED]                 = DIV_SUM,
+    [FMT_SEQUENCE]                 = DIV_SAME,
 };
 
 #undef CHANNEL