+++ src/subtract.c
@@ -35,12 +35,14 @@ typedef enum {
 
 static const sub_rule RULES[FMT_N_FIELDS] = {
     FMT_CHANNELS(CHANNEL)
-    [FMT_COVERAGE] = SUB_SUM,
-    [FMT_LENGTHS]  = SUB_SUM,
-    [FMT_READS]    = SUB_SUM,
-    [FMT_REJECTED] = SUB_SUM,
-    [FMT_UNMAPPED] = SUB_SUM,
-    [FMT_SEQUENCE] = SUB_SAME,
+    [FMT_COVERAGE]                 = SUB_SUM,
+    [FMT_LENGTHS]                  = SUB_SUM,
+    [FMT_PRIMARY_COUNTED]          = SUB_SUM,
+    [FMT_PRIMARY_REJECTED]         = SUB_SUM,
+    [FMT_SUPPLEMENTARY_COUNTED]    = SUB_SUM,
+    [FMT_SUPPLEMENTARY_REJECTED]   = SUB_SUM,
+    [FMT_UNMAPPED]                 = SUB_SUM,
+    [FMT_SEQUENCE]                 = SUB_SAME,
 };
 
 #undef CHANNEL